Update tests with downcast fixes.

+++ tests/test_sqlpool.py
@@ -256,7 +256,7 @@ def test_splitting_data_add_and_query():
     for i in range(1,2):
         poolfile.add_splitting_data(2*i, 1, 0.1, [2*i-1], [0], [0.0], [0.0, 0.0])
         poolfile.mark_last_iteration_as_completed()
-    assert np.all(poolfile.get_minmax()[0] == np.array([2.0,0.0,0.0], dtype="float32"))
+    assert np.all(poolfile.get_minmax()[0] == np.array([2.0,0.0,0.0], dtype="float64"))
 
 def test_splitting_data_add_update_and_query():
     """Adding splitting data to the database."""
@@ -264,7 +264,7 @@ def test_splitting_data_add_update_and_query():
     poolfile.add_splitting_data(2, 1, 0.1, [1], [0], [0.0], [0.0, 0.0])
     poolfile.update_splitting_data(2, 1, 0.1, [1], [0], [0.0], [0.0, 0.3])
     poolfile.mark_last_iteration_as_completed()
-    assert np.all(poolfile.get_minmax()[0] == np.array([2.0,0.0,0.3], dtype="float32"))
+    assert np.all(poolfile.get_minmax()[0] == np.array([2.0,0.0,0.3], dtype="float64"))
 
 @pytest.mark.usefixtures("skip_on_windows")
 def test_splitting_data_query_fail():
